fixed int cast
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@5633 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/libs/orcaice/timeutils.cpp
@@ -39,7 +39,7 @@ toOrcaTime( const IceUtil::Time & icetime )
     orca::Time orcatime;
     //! @todo are these casts ok?
     orcatime.seconds = (int)icetime.toSeconds();
-    orcatime.useconds = (int)icetime.toMicroSeconds() - (int)icetime.toSeconds()*1000000;
+    orcatime.useconds = (int)(icetime.toMicroSeconds() - icetime.toSeconds()*1000000);
     return orcatime;
 }
